requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
amqp==2.6.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
astroid==2.8.3; python_version ~= '3.6'
attrdict==2.0.1
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
authlib==1.0.0b1
billiard==3.6.4.0
boto3==1.9.170
botocore==1.12.253
celery-message-consumer==1.1.1
celery==4.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7; python_version >= '3'
configloader[attrdict,yaml]==1.0.1
coverage==4.5.2
cryptography==3.4.8; python_version >= '3.6'
dj-database-url==0.5.0
django-autofixture==0.12.1
django-cas-client==1.5.0
django-cors-headers==3.2.0
django-extensions==3.1.3
django-filter==2.1.0
django-log-request-id==1.7.0
django-nose==1.4.6
django-rest-knox==4.0.1
django==2.2.24
djangorestframework-expander==0.2.3
djangorestframework==3.12.4
docutils==0.15.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
entrypoints==0.3; python_version >= '2.7'
faker==9.5.2
flake8==3.7.6
flexisettings==1.0.5
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==4.8.1
iniconfig==1.1.1
isort==5.9.3; python_version < '4' and python_full_version >= '3.6.1'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
kombu==4.6.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
lazy-object-proxy==1.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
mccabe==0.6.1
nose==1.3.7
oauthlib==3.1.1; python_version >= '3.6'
packaging==21.0; python_version >= '3.6'
parameterized==0.7.0
pdfrw==0.4
pluggy==1.0.0; python_version >= '3.6'
psycopg2-binary==2.8.6
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycodestyle==2.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyflakes==2.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pylint-django==2.0.5
pylint-plugin-utils==0.6
pylint==2.2.2
pyparsing==3.0.0rc2; python_version >= '3.6'
pytest-cov==2.6.1
pytest==6.2.5; python_version >= '3.6'
python-coveralls==2.9.3
python-dateutil==2.8.2; python_version >= '2.7'
python-json-logger==2.0.2
pytz==2021.3
pyyaml==6.0
requests-oauthlib==1.3.0
requests==2.23.0
s3transfer==0.2.1
sentry-sdk==1.3.1
simple-salesforce==1.11.3
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
text-unidecode==1.3
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.2
typing==3.7.4.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.25.11; python_version >= '3.4'
vine==1.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
waitress==2.0.0
wrapt==1.13.3rc1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zipp==3.6.0; python_version >= '3.6'
